Method update and progress report
Confident that I am almost done with the askquestions method. Also have Unit testing almost done as well. I wonder if there is a way to check if the Test passed or fail and implement it instead of hard coding that the test passed. Like "Didi the test do what it needed to do?" "Yes? then test passed" "No? Test failed"